Add rudimentary tokenizer
Faced a lot of issues with the Token datatype and eventually resorted to Token keeping the actual string for the token instead of a string pointer or reference. There were a lot of compiler issues that have been documented.

To avoid the issue of expensive pass-by-value, either std::move or a move constructor needs to be implemented.

TODOs:

0. Improve tokenization (currently doesn't support lack of whitespace b/w common delimiters. e.g. procedure proc{x=5} is not supported because of the lack of white space (current way splits via space.
1. Add tagging logic
2. Syntax validator