Support basic syntax validations & fix minor bugs
Validation for Assignment statements not implemented yet.

Please note that the current implementations are very hard-codey. They need to be improved on to support recursive validation (e.g. for conditional expressions which may contain cond_exprs) of their own right.
Might need a discussion on that, don't have a vision for it right now.

Bugs fixed:
-  incorrect regex pattern led to mislabelling

Next Steps:
1. Helper functions like IsCondExpr() and such are needed for recursive calls