DE: Implement extract Modifies relationship
Same algorithm as Uses. Same procedures used for unit test.
Just realized that statements added to the procedures for
ExtractUses and ExtractModifies can be arbitrary if not a
container, but I'll just leave it first.

Idea considered: ExtractUses and ExtractModifies can be
concurrently using the same method

Idea rejected: The helper function in the recursion returns a
list of statements that are different for Uses and Modifies.
The smaller conditional helper functions cannot be used too
bcos the relationships would be double counted.