Finish logger implementation
Not sure how to make the output file be created outside of cmake-build-debug/
But I guess this is good enough? Haven't tested it across multiple functions but shouldn't be an issue since the logger is in global scope.
See the TestCommonUtils on a demo on how to use the LOG macro function