Improve utests for nested calls in ModifiesExtractor
Edited existing test cases for Modifies to reflect the change in psub: all local ModifiesP will be added in psub (except that for call stmts). This does not affect correctness.

The commented out code in ModifiesExtractor and UsesExtractor are meant to reduce repeated work of adding UsesP and ModifiesP relationships also done by PSubsystem, but cannot work until the call stmt can trace back to the current procedure.