Heavily refactor PSubsystem::HandleCloseBrace()
The closing of blocks and container logic is a little messy in general because of the unavoidable use of side-effects.
The previous implementation had too much code in one level, hence most of the function has been abstracted away and intuitive names have been added.
This refactoring ensures that every function has a similar level of abstration (SLAP hard 🍑 ) and should make it easier for me to add in the Cluster logic.

Existing test unit tests for CFG pass, but unsure if any breaking changes have been made, an eyeballing is definitely required.