FIx block creation issues and add ExitBlock creation
Block creation issue existed because of the use of the GetNextBlocks which was returning a set instead of a pointer, so adding to that set meant adding to a copy of that set actually,  so undesirable effects were seen.

Exit blocks are added after if and else blocks. so for an if cluster, we expect: if cond block --> if body block --> else body block --> exit block
Exit blocks are uniquely identified by the range which is -INT_MAX, -INT_MAX. There's a helper function to check if it's an exit block. Note that the exit block is NOT a singleton obj.