Add RuntimeMediator and Colleague interface
Adding Mediator into RuntimeColleague interface will cause a circular dependency that I cannot fix.  The intention was to make the extractors implement the colleague interface while requiring the mediator but it has the mediator dependency. The workaround can be to set the mediator as a method in the concrete extractors.